Fix find_path() parameter types to unsigned long

+++ src/bfs.h
@@ -1,5 +1,5 @@
 #pragma once
 #include "globals.h"
 
-void find_path(int start, int end, int parents[]);
+void find_path(unsigned long start, unsigned long end, unsigned long parents[]);
 void print_parents(graph *g);
